FarmId 발급 기능 추가
- FarmId Service 추가

+++ Temp_Spring/src/main/java/com/example/temp_spring/domain/data/FarmIdData.java
@@ -18,7 +18,7 @@ import javax.persistence.Id;
 public class FarmIdData {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // P
+    private Long id; // PK
 
-    private Long farmId;
+    private int farmId;
 }
